Version 3.0-alpha-1
create addon directory

divide files

divide files part 2

divide files part 3

complete: divide files

add feature: flip/rotate UVs
